pyproject.toml
[project]
name = "django-cookbook"
description = "Django Cookbook"
authors = [{ name = "Lucas", email = "lucassix.lee@gmail.com" }]
readme = "README.md"
requires-python = "~=3.12"
license-files = ["LICEN[CS]E*"]
maintainers = [{ name = "Lucas", email = "lucassix.lee@gmail.com" }]
keywords = ["django"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",

    "Intended Audience :: Developers",
    "Topic :: Software Development :: Documentation",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",

    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    "Operating System :: OS Independent",
    "Framework :: Django :: 4",
    "Framework :: Django :: 4.2",
    "Typing :: Typed",
]
dependencies = [
    "django~=4.2",
    "django-stubs[compatible-mypy]",
    "psycopg[binary, pool] >= 3.2",
    "redis[hiredis]",
    "types-redis",

    "djangorestframework",
    "djangorestframework-stubs",
    "django-filter",
    "django-filter-stubs",
    "django-crispy-forms",
    "crispy-bootstrap5",
]
dynamic = ["version"]

[project.urls]
Home = "https://lucas-six.github.io/django-cookbook/"
Documentation = "https://lucas-six.github.io/django-cookbook/"
Repository = "https://github.com/lucas-six/django-cookbook/"

[tool.setuptools]
py-modules = ['django_cookbook']

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true

[dependency-groups]
dev = ["mypy>=1.15.0", "colorlog", "django-debug-toolbar", "ruff>=0.11.8"]

[tool.ruff]
line-length = 100
lint.extend-safe-fixes = [
    # non-pep585-annotation
    "UP006",
]
lint.select = [
    # flake8-bugbear
    "B",
    # flake8-comprehensions
    "C4",
    # pycodestyle
    "E",
    # Pyflakes errors
    "F",
    # isort
    "I",
    # flake8-simplify
    "SIM",
    # flake8-tidy-imports
    "TID",
    # pyupgrade
    "UP",
    # Pyflakes warnings
    "W",
]
lint.ignore = []

[tool.ruff.format]
quote-style = "single"

[tool.mypy]
python_version = "3.12"
exclude = ["manage.py", "models.py", "migrations/"]
follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
warn_unused_configs = true
disallow_any_generics = false
check_untyped_defs = true
no_implicit_reexport = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "redis.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "debug_toolbar.*"
ignore_missing_imports = true

[tool.pyright]
include = ["django_cookbook"]
exclude = [
    ".git",
    "**/__pycache__",
    ".venv",
    ".mypy_cache",
    "*.egg-info",
    "**/migrations",
]
reportIncompatibleVariableOverride = "none"
reportGeneralTypeIssues = "none"
stubPath = ""
pythonVersion = "3.12"
